refactor(ui): move Storm Sigils to dedicated stats panel section
Extract sigil rendering from Equipment sub-panel into its own
stats_sigils.rs module. Sigil panel appears between Attributes
and Equipment when at least 1 sigil is etched, with dynamic height.
Equipment block is now purely equipment again.
